Polyfill for get_course_and_cm_from_instance()
Introduced in MDL-46706 and merged in Moodle 2.8, this enhancement
provides an easy means for developers to obtain modinfo and cm_info
objects from DML objects representing module records.

Unfortunately, the use of this feature breaks the Behat suite for
releases of Moodle prior to 2.8, despite the plugin also supporting Moodle
versions 2.0 through 2.7.